Refactor tutor code + save observability logs
